--- lib/src/components/data-table/bulk-actions/components/index.ts
@@ -1,2 +1,0 @@
-export { BulkActionsDefaultActions } from './BulkActionsDefaultActions'
-export { BulkActionsSelectedRowActions } from './BulkActionsSelectedRowActions'
