--- lib/src/components/data-table/bulk-actions/index.ts
@@ -1,1 +1,0 @@
-export { DataTableBulkActions } from './DataTableBulkActions'
